# Remove unused header file