--- a/include/util/string_util.hpp
+++ b/include/util/string_util.hpp
@@ -10,53 +10,6 @@
 namespace osrm::util
 {
 
-// precision:  position after decimal point
-// length: maximum number of digits including comma and decimals
-// work with negative values to prevent overflowing when taking -value
-template <int length, int precision> char *printInt(char *buffer, int value)
-{
-    static_assert(length > 0, "length must be positive");
-    static_assert(precision > 0, "precision must be positive");
-
-    const bool minus = [&value]
-    {
-        if (value >= 0)
-        {
-            value = -value;
-            return false;
-        }
-        return true;
-    }();
-
-    buffer += length - 1;
-    for (int i = 0; i < precision; ++i)
-    {
-        *buffer = '0' - (value % 10);
-        value /= 10;
-        --buffer;
-    }
-    *buffer = '.';
-    --buffer;
-
-    for (int i = precision + 1; i < length; ++i)
-    {
-        *buffer = '0' - (value % 10);
-        value /= 10;
-        if (value == 0)
-        {
-            break;
-        }
-        --buffer;
-    }
-
-    if (minus)
-    {
-        --buffer;
-        *buffer = '-';
-    }
-    return buffer;
-}
-
 inline bool RequiresJSONStringEscaping(const std::string &string)
 {
     for (const char letter : string)