--- a/unit_tests/util/string_util.cpp
+++ b/unit_tests/util/string_util.cpp
@@ -2,7 +2,7 @@
 
 #include <boost/test/unit_test.hpp>
 
-#include <iostream>
+#include <string>
 
 BOOST_AUTO_TEST_SUITE(string_util)
 
@@ -25,22 +25,6 @@ BOOST_AUTO_TEST_CASE(json_escaping)
     BOOST_CHECK_EQUAL(output, "Aleja \\\"Solidarnosci\\\"");
 
     BOOST_CHECK(!RequiresJSONStringEscaping("Aleja Solidarnosci"));
-}
-
-BOOST_AUTO_TEST_CASE(print_int)
-{
-    const std::string input{"\b\\"};
-    char buffer[12];
-    buffer[11] = 0; // zero termination
-    std::string output = printInt<11, 8>(buffer, 314158976);
-    BOOST_CHECK_EQUAL(output, "3.14158976");
-
-    buffer[11] = 0;
-    output = printInt<11, 8>(buffer, 0);
-    BOOST_CHECK_EQUAL(output, "0.00000000");
-
-    output = printInt<11, 8>(buffer, -314158976);
-    BOOST_CHECK_EQUAL(output, "-3.14158976");
 }
 
 BOOST_AUTO_TEST_SUITE_END()