<!-- BENCHMARK_RESULTS_START -->
## Benchmark Results
| Benchmark | Base | PR |
|-----------|------|----|
| alias | aliased u32: 1100.96<br/>plain u32: 1090.33<br/>aliased double: 959.596<br/>plain double: 952.866 | aliased u32: 1097.05<br/>plain u32: 1083.09<br/>aliased double: 952.836<br/>plain double: 956.151 |
| json-render | String: 6.5741ms<br/>Stringstream: 9.3445ms<br/>Vector: 6.8484ms | String: 6.59736ms<br/>Stringstream: 9.3328ms<br/>Vector: 6.87642ms |
| match_ch | Default radius: <br/>4.51663ms/req at 82 coordinate<br/>0.0550809ms/coordinate<br/>Radius 5m: <br/>4.49462ms/req at 82 coordinate<br/>0.0548124ms/coordinate<br/>Radius 10m: <br/>15.4135ms/req at 82 coordinate<br/>0.18797ms/coordinate<br/>Radius 15m: <br/>37.9857ms/req at 82 coordinate<br/>0.463241ms/coordinate<br/>Radius 30m: <br/>321.845ms/req at 82 coordinate<br/>3.92494ms/coordinate | Default radius: <br/>4.47706ms/req at 82 coordinate<br/>0.0545983ms/coordinate<br/>Radius 5m: <br/>4.44987ms/req at 82 coordinate<br/>0.0542667ms/coordinate<br/>Radius 10m: <br/>15.3191ms/req at 82 coordinate<br/>0.186818ms/coordinate<br/>Radius 15m: <br/>37.3968ms/req at 82 coordinate<br/>0.456059ms/coordinate<br/>Radius 30m: <br/>318.964ms/req at 82 coordinate<br/>3.88981ms/coordinate |
| match_mld | Default radius: <br/>3.01219ms/req at 82 coordinate<br/>0.036734ms/coordinate<br/>Radius 5m: <br/>2.90696ms/req at 82 coordinate<br/>0.0354507ms/coordinate<br/>Radius 10m: <br/>10.4022ms/req at 82 coordinate<br/>0.126856ms/coordinate<br/>Radius 15m: <br/>26.7163ms/req at 82 coordinate<br/>0.325809ms/coordinate<br/>Radius 30m: <br/>312.853ms/req at 82 coordinate<br/>3.81528ms/coordinate | Default radius: <br/>3.02108ms/req at 82 coordinate<br/>0.0368425ms/coordinate<br/>Radius 5m: <br/>2.9898ms/req at 82 coordinate<br/>0.0364609ms/coordinate<br/>Radius 10m: <br/>11.5109ms/req at 82 coordinate<br/>0.140377ms/coordinate<br/>Radius 15m: <br/>29.9391ms/req at 82 coordinate<br/>0.365112ms/coordinate<br/>Radius 30m: <br/>357.829ms/req at 82 coordinate<br/>4.36377ms/coordinate |
| packedvector | random write:<br/>std::vector 9859.32 ms<br/>util::packed_vector 78207 ms<br/>slowdown: 7.93229<br/>random read:<br/>std::vector 8509.16 ms<br/>util::packed_vector 33496.2 ms<br/>slowdown: 3.93649 | random write:<br/>std::vector 9824.27 ms<br/>util::packed_vector 73791.5 ms<br/>slowdown: 7.51114<br/>random read:<br/>std::vector 8480.41 ms<br/>util::packed_vector 30783.3 ms<br/>slowdown: 3.62993 |
| route_ch | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>511.938ms<br/>0.511938ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>354.195ms<br/>0.354195ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>634.456ms<br/>0.634456ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>151.942ms<br/>0.151942ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>98.3052ms<br/>0.0983052ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>135.258ms<br/>0.135258ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>151.411ms<br/>0.151411ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>98.6195ms<br/>0.0986195ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>135.069ms<br/>0.135069ms/req | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>510.432ms<br/>0.510432ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>351.994ms<br/>0.351994ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>627.408ms<br/>0.627408ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>152.496ms<br/>0.152496ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>98.0008ms<br/>0.0980008ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>132.527ms<br/>0.132527ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>151.009ms<br/>0.151009ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>97.8028ms<br/>0.0978028ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>131.42ms<br/>0.13142ms/req |
| route_mld | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>644.595ms<br/>0.644595ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>443.93ms<br/>0.44393ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>821.843ms<br/>0.821843ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>288.7ms<br/>0.2887ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>166.365ms<br/>0.166365ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>305.72ms<br/>0.30572ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>289.509ms<br/>0.289509ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>167.357ms<br/>0.167357ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>306.084ms<br/>0.306084ms/req | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>654.292ms<br/>0.654292ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>441.942ms<br/>0.441942ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>839.873ms<br/>0.839873ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>298.655ms<br/>0.298655ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>165.67ms<br/>0.16567ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>315.596ms<br/>0.315596ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>300.174ms<br/>0.300174ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>164.949ms<br/>0.164949ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>313.095ms<br/>0.313095ms/req |
| rtree | 1 result:<br/>208.856ms ->  0.0208856 ms/query<br/>10 results:<br/>243.617ms ->  0.0243617 ms/query | 1 result:<br/>207.746ms ->  0.0207746 ms/query<br/>10 results:<br/>242.442ms ->  0.0242442 ms/query |
<!-- BENCHMARK_RESULTS_END -->